backend/requirements.txt
# --- Core Backend Framework ---
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
requests==2.31.0
httpx==0.24.1
python-dotenv==1.0.0

# --- Database & Storage ---
supabase==2.0.2
redis==5.0.1

# --- AI / NLP / ML ---
sentence-transformers==2.2.2
transformers==4.36.2
torch==2.1.2
scikit-learn==1.3.2
numpy==1.24.3

# --- Image Processing ---
Pillow==10.1.0

# --- Testing ---
pytest==7.4.3
